Download dependencies before comparing parse reults
When local Maven repo is empty the Markers added by the RewriteMavenProjectParser (comparing) and the RewriteProjectParser (under test) differ.
This is because one downloads the dependencies and the next takes them form local Maven and the URIs for dependencies then differ.
By builing the projects with Maven first the dependencies will be available in local Maven repo and taken from there by both parsers.